requirements.txt
# Core PDF Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pypdf>=3.17.0

# AI Enhancement
google-generativeai>=0.3.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# NER and ML
spacy>=3.7.0
sentence-transformers>=2.2.0

# Text Processing
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0

# Schema Validation
jsonschema>=4.19.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.7.0
